+++ src/components/Docs/data/continuousIntegrationData.ts
@@ -74,7 +74,7 @@ export const featuredTutorials: CardItem[] = [
           module: MODULES.ci,
           description:
             "Run builds at scale on VMs that are preconfigured with tools, packages, and settings commonly used in CI pipelines.",
-          link: "/docs/continuous-integration/ci-quickstarts/hosted-builds-on-virtual-machines-quickstart",
+          link: "/docs/continuous-integration/use-ci/set-up-build-infrastructure/use-harness-cloud-build-infrastructure",
         },
         {
           title: "Test Intelligence",